chore: merge repo & doc-engine into one before cleaning up
