+++ packages/core/core/src/services/document-service/common.ts
@@ -1,6 +1,6 @@
-import type { Common } from '@strapi/types';
+import type { Common, Documents } from '@strapi/types';
 
-export type RepositoryFactoryMethod = (uid: Common.UID.CollectionType) => any;
+export type RepositoryFactoryMethod = (uid: Common.UID.CollectionType) => Documents.ServiceInstance;
 
 export const wrapInTransaction = (fn: (...args: any) => any) => {
   return (...args: any[]) => strapi.db.transaction?.(() => fn(...args));
