+++ packages/core/types/src/modules/documents/service-instance.ts
@@ -1,5 +1,5 @@
-import type { Common, Documents } from '../..';
-import type { ID } from './document-engine';
+import type { Common } from '../..';
+import type { ID } from '.';
 import type * as Params from './params/document-engine';
 import type * as Result from './result/document-enigne';
 
@@ -54,30 +54,4 @@ export type ServiceInstance<
     documentId: ID,
     params?: TParams
   ) => Result.DiscardDraft<TContentTypeUID, TParams>;
-};
-
-export type SingleTypeInstance<
-  TContentTypeUID extends Common.UID.ContentType = Common.UID.ContentType
-> = {
-  find: <TParams extends Params.FindOne<TContentTypeUID>>(
-    params?: TParams
-  ) => Promise<Documents.Result<TContentTypeUID, TParams>>;
-
-  delete: <TParams extends Params.Delete<TContentTypeUID>>(params?: TParams) => Result.Delete;
-
-  update: <TParams extends Params.Update<TContentTypeUID>>(
-    params: TParams
-  ) => Promise<Documents.Result<TContentTypeUID, TParams>>;
-
-  publish: <TParams extends Params.Publish<TContentTypeUID>>(
-    params?: TParams
-  ) => Promise<Documents.Result<TContentTypeUID, TParams>>;
-
-  unpublish: <TParams extends Params.Unpublish<TContentTypeUID>>(
-    params?: TParams
-  ) => Promise<Documents.Result<TContentTypeUID, TParams>>;
-
-  discardDraft: <TParams extends Params.DiscardDraft<TContentTypeUID>>(
-    params?: TParams
-  ) => Promise<Documents.Result<TContentTypeUID, TParams>>;
 };
